change int to size_t

+++ src/staline.c
@@ -8,8 +8,8 @@ int* i_staline_sort_copy(int* arr, size_t* length, int starting_index) {
     int* temp = malloc(*length * sizeof(int));
     temp[0] = arr[starting_index];
 
-    int current = 0;
-    for (int i = starting_index + 1; i < *length; i++) {
+    size_t current = 0;
+    for (size_t i = starting_index + 1; i < *length; i++) {
         if (arr[i] >= temp[current]) {
             temp[++current] = arr[i];
         }
@@ -20,8 +20,8 @@ int* i_staline_sort_copy(int* arr, size_t* length, int starting_index) {
 }
 
 void staline_sort_in_place(int* arr, size_t* length) {
-    int current = 0;
-    for (int i = 1; i < *length; i++) {
+    size_t current = 0;
+    for (size_t i = 1; i < *length; i++) {
         if (arr[i] >= arr[current] && ++current != i) {
             // arithmetic swap
             arr[current] = arr[current] + arr[i];
@@ -36,7 +36,7 @@ int* biggest_staline_sort(int* array, size_t* length) {
     int* result = NULL;
     size_t result_length = 0;
 
-    for (int i = 0; i < *length - 1; i++) {
+    for (size_t i = 0; i < *length - 1; i++) {
         size_t current_length = *length;
         int* current_array = i_staline_sort_copy(array, &current_length, i);
         if (current_length > result_length) {
